move commit to licAppRepo
