+++ src/Spd.Resource.Repository.IntegrationTest/PersonLicApplicationRepositoryTest.cs
@@ -2,10 +2,9 @@ using AutoFixture;
 using Microsoft.Dynamics.CRM;
 using Microsoft.Extensions.DependencyInjection;
 using Spd.Resource.Repository.Alias;
-using Spd.Resource.Repository.Application;
+using Spd.Resource.Repository.LicApp;
 using Spd.Resource.Repository.PersonLicApplication;
 using Spd.Utilities.Dynamics;
-using Spd.Utilities.Shared.Exceptions;
 
 namespace Spd.Resource.Repository.IntegrationTest;
 
@@ -85,59 +84,5 @@ public class PersonLicApplicationRepositoryTest : IClassFixture<IntegrationTestS
     {
         // Action and Assert
         await Assert.ThrowsAsync<ArgumentException>(async () => await _personLicAppRepository.GetLicenceApplicationAsync(Guid.NewGuid(), CancellationToken.None));
-    }
-
-    [Fact]
-    public async Task CommitLicenceApplicationAsync_ForApplicant_Run_Correctly()
-    {
-        // Arrange
-        Guid appId = Guid.NewGuid();
-        Guid contactId = Guid.NewGuid();
-        spd_application app = new() { spd_applicationid = appId };
-        _context.AddTospd_applications(app);
-        contact contact = new() { contactid = contactId };
-        _context.AddTocontacts(contact);
-        _context.SetLink(app, nameof(spd_application.spd_ApplicantId_contact), contact);
-        await _context.SaveChangesAsync();
-
-        // Action
-        LicenceApplicationCmdResp? resp = await _personLicAppRepository.CommitLicenceApplicationAsync(appId, ApplicationStatusEnum.Submitted, CancellationToken.None);
-
-        //Assert
-        Assert.NotNull(resp);
-        Assert.Equal(appId, resp.LicenceAppId);
-        Assert.Equal(contactId, resp.ContactId);
-
-        // Annihilate
-        _context.DeleteObject(app);
-        _context.DeleteObject(contact);
-        await _context.SaveChangesAsync(CancellationToken.None);
-    }
-
-    [Fact]
-    public async Task CommitLicenceApplicationAsync_ForBusiness_Run_Correctly()
-    {
-        // Arrange
-        Guid appId = Guid.NewGuid();
-        Guid bizId = Guid.NewGuid();
-        spd_application app = new() { spd_applicationid = appId };
-        _context.AddTospd_applications(app);
-        account account = new() { accountid = bizId };
-        _context.AddToaccounts(account);
-        _context.SetLink(app, nameof(spd_application.spd_OrganizationId), account);
-        await _context.SaveChangesAsync();
-
-        // Action
-        LicenceApplicationCmdResp? resp = await _personLicAppRepository.CommitLicenceApplicationAsync(appId, ApplicationStatusEnum.Submitted, CancellationToken.None);
-
-        //Assert
-        Assert.NotNull(resp);
-        Assert.Equal(appId, resp.LicenceAppId);
-        Assert.Equal(bizId, resp.ContactId);
-
-        // Annihilate
-        _context.DeleteObject(app);
-        _context.DeleteObject(account);
-        await _context.SaveChangesAsync(CancellationToken.None);
     }
 }
