+++ src/Spd.Resource.Repository/LicApp/Contract.cs
@@ -5,6 +5,8 @@ namespace Spd.Resource.Repository.LicApp;
 public partial interface ILicAppRepository
 {
     public Task<IEnumerable<LicenceAppListResp>> QueryAsync(LicenceAppQuery qry, CancellationToken cancellationToken);
+    //connect spd_application with spd_contact and update application to correct status
+    public Task<LicenceApplicationCmdResp> CommitLicenceApplicationAsync(Guid applicationId, ApplicationStatusEnum status, CancellationToken ct);
 }
 
 public record LicenceAppQuery(Guid? ApplicantId, Guid? BizId, List<WorkerLicenceTypeEnum>? ValidWorkerLicenceTypeCodes, List<ApplicationPortalStatusEnum>? ValidPortalStatus);
@@ -21,6 +23,7 @@ public record LicenceAppListResp
     public ApplicationPortalStatusEnum ApplicationPortalStatusCode { get; set; }
 }
 
+public record LicenceApplicationCmdResp(Guid LicenceAppId, Guid? ContactId = null, Guid? BizId = null);
