+++ src/Spd.Resource.Repository/LicApp/LicAppRepository.cs
@@ -1,6 +1,9 @@
 using AutoMapper;
 using Microsoft.Dynamics.CRM;
+using Spd.Resource.Repository.Application;
 using Spd.Utilities.Dynamics;
+using Spd.Utilities.Shared.Exceptions;
+using System.Net;
 
 namespace Spd.Resource.Repository.LicApp;
 internal class LicAppRepository : ILicAppRepository
@@ -14,6 +17,34 @@ internal class LicAppRepository : ILicAppRepository
         _mapper = mapper;
     }
 
+    //for unauth, set applcation status to submitted.
+    public async Task<LicenceApplicationCmdResp> CommitLicenceApplicationAsync(Guid applicationId, ApplicationStatusEnum status, CancellationToken ct)
+    {
+        spd_application? app = await _context.GetApplicationById(applicationId, ct);
+        if (app == null)
+            throw new ApiException(HttpStatusCode.BadRequest, "Invalid ApplicationId");
+
+        if (status == ApplicationStatusEnum.Submitted)
+        {
+            app.statuscode = (int)ApplicationStatusOptionSet.Submitted;
+            app.statecode = DynamicsConstants.StateCode_Inactive;
+        }
+        else
+        {
+            app.statuscode = (int)Enum.Parse<ApplicationStatusOptionSet>(status.ToString());
+        }
+
+        app.spd_submittedon = DateTimeOffset.Now;
+        app.spd_portalmodifiedon = DateTimeOffset.Now;
+        _context.UpdateObject(app);
+        await _context.SaveChangesAsync(ct);
+
+        // For business application, return organization id, for all others, return applicant id
+        if (app._spd_organizationid_value != null)
+            return new LicenceApplicationCmdResp((Guid)app.spd_applicationid, null, (Guid)app._spd_organizationid_value);
+        else
+            return new LicenceApplicationCmdResp((Guid)app.spd_applicationid, (Guid)app._spd_applicantid_value, null);
+    }
 
     public async Task<IEnumerable<LicenceAppListResp>> QueryAsync(LicenceAppQuery qry, CancellationToken cancellationToken)
     {
